Added Escape key for Input
Added a low level keyboard hook to stop the escape key from being
pressed and allow it to be used for key input  mapping such as Xbox
Logo/Guide. (DOES NOT WORK ON 32 BIT MACHINES, Yet?)

For a simpler build process:
Setup a simple-ish bat script to automatically build the files with
names (automatically builds both a 32bit and 64bit version of the
application - This is needed because of the Escape hook only works on 64
bit machines.)

Move around classes in folders for a cleaner experience when viewing
code.